mgr/cephadm: move `forall_hosts()` to utils module
